temperature/requirements.txt
cryptography==36.0.1
google-api-python-client==2.34.0
google-auth-httplib2==0.1.0
google-auth==2.3.3
google-cloud-pubsub==2.9.0
google-cloud-iot==2.3.0
grpc-google-iam-v1==0.12.3
pyjwt==2.3.0
paho-mqtt==1.5.1
pyyaml
